clean-ups and test adjustments
